27869: Lint rule for LinearLayout which doesn't specify orientation
This adds a check for cases where you have a LinearLayout which
does not specify an orientation (meaning it's horizontal), yet
it contains multiple children, and one of the children prior to
the last one sets its width to fill_parent or match_parent (without
a weight).

This is a common pitfall for new developers who create the layout
and assume it's vertical, and at runtime only see the first child,
not realizing the remainder are off screen to the right.

Change-Id: Ief9604d8a8fc84849d1a8f2df6d7ec0b75b13eaa