Make resource folders customizeable
Until now, lint has hardcoded where the project resources are to be
found: in $project/res.

With the new build system this will no longer necessarily be the case;
there can be multiple resource folders, and the locations might not be
just res/.

Therefore, this CL generalizes lint's handling of resource folders: it
is now provided by the LintClient, and the command line lint tool uses
this to offer a new --resources flag which can be used to set the
resource directory/directories for a project.

Change-Id: I3717c6474141776ee5541d476f9bad5bb8e1a9c7