hv: Refactor to interact with the device via IHvDevice
Hierarchy Viewer currently interacts with a View Server that
is present on the device. This is available only on eng devices
due to security restrictions on the View Server.

Rather than use this custom view server, we ought to use DDM
for communication with the device. Such a scheme has a number of
benefits apart from security.

This CL is primarily just a refactoring of the existing host
side code. The main objective is to hide the communication to
the device behind a IHvDevice interface. The ViewServerDevice
implementation of this interface allows communicating with
existing devices that use the ViewServer implementation on
the device. A subsequent CL will provide a new implementation
of this interface that communicates via DDM.

Change-Id: I7d63e5a59c6ec9c96dbd07af9dc03f93779fd2ec