Support modifying layout properties from hv
This CL adds support in the hierarchy viewer's UI for modifying
certain view properties. The editable properties are marked with
an image next to them. When edited, a command is sent to the
device indicating the view method to call and the arguments for it.

A free form text box is also added to the UI which allows for users
to explicitly specify the view method to call with parameters.

Change-Id: Iefd4239e91184fc43c1b4cccfc3896846ab19825