Extended the MonkeyRunner shell method with timeout arg.
In python via MonkeyRunner it is possible execute shell
commands but the running time was reduced to 5 seconds.
This patch adds an extra argument for the "shell" method.
It decides the maximum amount of time during which the
command is allowed to not output any response. A value of
0 means that the method will wait forever. This extra arg
is optionally. So the original function works without any changes.

Change-Id: I1a6d9b37e1846715342b252b2a2a346c2309a08f
Signed-off-by: Zoltán Papp <pzatya@gmail.com>