Fix API check via support library
This CL fixes the API check such that it properly handles calls
in the Android support library (android.support.*). There was an
optimization to skip looking up inherited API calls once the API
checker reaches the android.* namespace, since at that point all
the inherited methods are inlined in the database, but this should
not be done if the package starts with android.support.*, since
those methods still could be inheriting from the Android APIs
without appearing in the database.

Change-Id: Iab7b9f98daa5507794a0673a8db53ce953f2bcd9