Allow editing patch regions via dragging markers
Patches are created in draw9patch by clicking on the 1 pixel
boundary. This is quite cumbersome when creating patches
that span certain regions.

This CL adds support for dragging markers to mark ranges
that are stretchable (or indicate content region). Based on
the location of the mouse pointer, a patch or padding region
is created in the image.

Change-Id: Ic0203efe9b12d4616f1eb76ddd3ca55d0f81cf01