Misc bug fixes
First, when computing the fully qualified package names of activities
to store the activity to theme mapping in the ManifestInfo, also allow
activities to not be prefixed by a dot and still prefix the package
path provided there are no dots in the activity name.

Second, escape string values specified in the Add Translation Dialog
when creating the corresponding resources.

Third, soon the ResourceRepository will no longer provide a protected
mWrapper field, so replicate it inside ProjectResources.

Change-Id: I7660aa6921e3b667e63a25c645ca21efc443b7ce