Modified templates for Gradle projects
Modifies the existing templates to output Gradle build files with a
new multi-level directory structure. The templates continue to work
the old way that Eclipse expects, and the new behavior is only enabled
if the "isGradle" global is set to "true".

Files to configure the Gradle wrapper are placed in the SDK at
tools/templates/gradle/wrapper.

This CL changes all file copies and merges to go to fully-qualified
directories set up via global variables. Things go to the old
locations for non-Gradle builds based on how the globals are set up,
but Gradle builds can override output directory variables to reflect
the new directory structure.

This CL also adds NewAndroidLibrary and NewJavaLibrary project
templates, which will be used in the future. The current
NewAndroidApplication template also can output a library, which is
redundant and confusing; this will likely be fixed somehow in the
future.

Change-Id: Ieb830bdfba1797d87a218326ebd37c99d254758e