Improves Gradle export.
Splits figuring out the modules and their props from
the code writing the new files.

The project selection page now uses a ProjectSetupBuilder
and updates it on every project selection or unselection.
This provides real-time feedback on missing dependencies
or broken ones.

Creates a new page to allow confirmation of project
root and list of modules. Will also allow review errors
when some are detected during modules gathering.

A final page, is shown after the export takes place
to review the result.

Change-Id: Ia67acc88f390f661518ed13818f25750be181c78