Check the return value of eglChooseConfig
The EGL specs say that eglChooseConfig doesn't update num_config
if it returns a failure (which is exactly what
Translator/EGL/EglImp.cpp does). Therefore, if this function
returned a failure (e.g. due to an unsupported egl attribute),
nConfigs was left untouched, meaning that the configs array
was left uninitialized but treated as if it was full of valid
configs.

Change-Id: I3809272298ea10d65dc939849d2e3c17d1158da6