Fixed ADT plugin bug while import existing project
When tries to import existing project whose name is dupliacated with the one
which is already in workspace, an error occurred with 'Cannot import [Project_
name] because the project name is in use.

But after user changes the project name to the other one which is not duplicated,
it just updates the project name on the list without updating project validity
which results still shows error message due to project name conflict in workspace.

This patch refreshes project name validity when user changes the project name
on the list.

Change-Id: Ia15c08767f0187061f2049a06e0b57051e5c767c
Signed-off-by: Taeho Kim <jyte82@gmail.com>