Allow control over whether Device Chooser Dialog is displayed
This CL allows users to control whether the device chooser dialog
will be displayed during a launch.

Currently, when performing a launch, the chooser dialog is displayed.
It has a checkbox for users to indicate that the same selected device
should be used for further launches. However, once set, there was no
way that the setting could be overridden to show the dialog again.

This CL mirrors that checkbox in the launch configuration. This allows
users to modify the launch configuration and disable the "Use last
selected device" checkbox, thereby forcing the chooser dialog to be
displayed again.

Fixes https://code.google.com/p/android/issues/detail?id=56076

Change-Id: I19dce8dd3acb76e799a422568267f28318b5ef58