emulator/opengl: Refactor SmartPtr implementation.
This gets rid of two copies of SmartPtr.h and replaces them with
a single implementation under shared/emugl/common/smart_ptr.*

Note that this uses a new include path rooted at the shared/
directory for classes that are likely to be built both for
the host and the device (in case we back-port this to
device/generic/goldfish/opengl/ in the future).

+ Add a gtest-based set of unittests, after building, just
  call 'emugl_common_host_unittests' to run it.

  Note that this probably needs a 64-bit version as well,
  will come later once I find a way to build GTest for 64-bits
  without breaking the platform build :-)

Also note that this moves the class to the 'emugl' namespace,
in order to make the code easier to build out of the platform
tree, and embed it in other projects. More classes will be
transitioned / refactored in future patches.

AOSP_BUG=64806

Change-Id: Ieb326c5f3f002a21537b8a391a82ce2ef9925073