emulator: Remove obsolete qtools directory.
The executable declared in this directory were used to process traces
that are generated with the "emulator -trace <name>" option. Unfortunately
this feature is broken in many ways:

- Only works for ARMv5, not ARMv7, x86 and MIPS
- Generates invalid timings that do not reflect modern CPU
  architectures.

Nobody has been using it for a long time, and the feature will be
removed from external/qemu, including the removal of one header file
included by the code under sdk/emulator/qtools.

This patch simply removes the qtools sources entirely.

Change-Id: I6e7d141393e71afb36d950c264f73752d42dc412