emulator/opengl: Remove <utils/List.h> + <utils/String8.h>
This removes dependencies on two libutils headers.

Rather than provide alternative data types, simply disable the
corresponding code because it looks like the corresponding fields
are simply never used by the code?

It is likely that this code is only used in the guest portion
of the code, and thus completely un-necessary on the host side.

Change-Id: I22c67f114c2c77381ceba8de23dd631158276596