Cherry pick from AOSP to fix build; please do not merge
Remove monitor temporarily from sdk builds (try 2)

This CL disables building of monitor from the SDK builds. This is
because the current PDE build doesn't work with Java7 on Macs. The
fix is to move away from PDE builds to Tycho, and that is going to
take a while. So we disable monitor until we move to Tycho.

Change-Id: Ibd701dc0fbd9b57441fa886fe701dcb71764e83a