Add hex prefix to systrace tag
For the V1 options dialog, Integer.toHexString is used to
write the value of the debug.atrace.tags.enableflags system
property, but it lacks the "0x" prefix. When this value is later
read by the framework with strtoll, it incorrectly interprets
the number as base 10 instead of hex.

This change updates the writing code to include a "0x" prefix
before the hex value, thereby ensuring that it is correctly
interpreted as a hex value in subsequent read operations.

Change-Id: I3fdb2f99d7e89dd197ee5433c2f54b14f147b62c