ADT: fix spurious SDK refresh when opening layout editor.
SDK Bug: b.android.com/67084

Reason: we recently switched SDK Manager location to a File
instead of string. ADT tries to automatically refresh the
SDK when it has changed or its path has changed and the
comparison between a File-to-path and the original path
string from the pref now differed by an ending /.

Also renames the legacy method to avoid confusion.

Change-Id: Ie4349075528fb552efe10cac4b4462b4bfd3a13f