Include all dependencies in the p2 repository
Currently, the generated p2 repository only includes the ADT
plugins. We need to include all the tools and prebuilts that
are now OSGI bundles, so this CL instructs p2 to include all
the dependencies as well.

Unfortunately, this means that all of the platform, jdt, cdt
are all included as well.

Change-Id: I5d2bfc8f72b18145e60985bffb54074a83a7d26a