Revert back from OSGI -> bundled jars
We tried the approach of replacing bundled jars with OSGI
jars. However, this eventually ended up resulting in
ClassCircularityError thrown at runtime when running on
Java 7. As far as we could diagnose, the issue seems to
be that:
 - LayoutLibrary attempts to load layoutlib.jar (which is
   a regular jar present inside the android sdk data folder).
 - The class loader passed to layoutlib is the LayoutLibrary's
   class loader.
 - The classes inside the layout library then attempt to
   load classes from a different OSGI jar (layoutlib-api)
 - At this time, we receive a ClassCircularityError

Most of our other jars depend on layoutlib-api, so it is
not possible to just bundle layoutlib-api. So for now, we
are back to bundling jars.

Change-Id: Ic8c838a1a7e90e1bc6b8c5251341b1ded46053ba